+++ CHANGELOG.md
@@ -37,6 +37,7 @@ All notable changes to `homebridge-config-ui-x` will be documented in this file.
 - move lang, theme, temp units + port to main settings
   - relabelled the UI settings screen as 'UI Advanced Settings'
 - offer to unpair child bridges when uninstalling plugins
+- offer to set up a child bridge immediately after configuring a new plugin
 - updates to `pl` language file (#2181) (@mkz212)
 - updates to `th` language file (#2182) (@tomzt)